+++ changelog.md
@@ -1,4 +1,7 @@
-# SharePoint Online Portal Showcase Changelog #
+# SharePoint Starter Kit Changelog #
+
+**2018-06-14**
+* Release blog post out in Office Dev blog
 
 **2018-02-27**
 * Created repository. It's GO!